Fix bug that sometimes we scrolled to the very bottom of list
The root cause is that list of entry objects and list of element
references do not have to be in sync - using one index for both does not
guarantee that we will work with DOM element that is created based on
underlying data. I think this is caused by Vue updating existing DOM
elements instead of creating whole list anew.